server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
alembic==1.13.0
pytest==7.4.3
httpx==0.25.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
# MySQL driver
pymysql==1.1.0
# AWS libraries
boto3==1.34.5
botocore==1.34.5
awscli==1.32.5
